// backend/src/constants/enums.ts
export enum ErrorCodeEnum {
  BAD_REQUEST = 400000,
  INVALID_PARAMETER = 400001,
  MISSING_EMAIL_CLIENT_CREDS = 400002,
  UNIQUE_CONSTRAINT = 400003,
  NOT_FOUND = 404000,
  UNAUTHORIZED = 401000,
  TOKEN_EXPIRED = 401001,
  INVALID_TOKEN = 401002,
  INCORRECT_ANSWER = 401003,
  EXPIRED_TOKEN = 401004,
  PASSWORD_MISMATCH = 401005,
  ACCOUNT_UNVERIFIED = 401006,
  INTERNAL_SERVER_ERROR = 500000,
}

// export enum FileContentTypeEnum {
//   PDF = "application/pdf",
//   CSV = "text/csv",
//   XLSX = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
//   JSON = "application/json",
//   ZIP = "application/zip",
//   PNG = "image/png",
//   JPEG = "image/jpeg",
//   JPG = "image/jpg",
//   WEBP = "image/webp",
//   AVI = "video/x-msvideo",
//   SVG = "image/svg+xml",
//   HTML = "text/html",
//   TEXT = "text/plain",
//   XML = "application/xml",
//   MP4 = "video/mp4",
//   MOV = "video/quicktime",
//   MP3 = "audio/mpeg",
//   WEBM = "video/webm",
//   OCTET_STREAM = "application/octet-stream",
//   XLS = "application/vnd.ms-excel",
// }